requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
backoff==2.2.1
billiard==4.2.4
celery==5.6.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
clerk-backend-api==5.0.6
clerk-sdk==1.1.0
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.7
debugpy==1.8.20
Django==6.0.3
django_celery_results==2.6.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
ecdsa==0.19.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
kombu==5.6.2
markdown-it-py==4.0.0
mdurl==0.1.2
packaging==26.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.2
python-jose==3.5.0
pytz==2026.1.post1
redis==7.4.0
requests==2.33.1
rich==14.3.3
rsa==4.9.1
six==1.17.0
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
